all done

+++ src/main/java/org/example/Statistics.java
@@ -2,17 +2,17 @@ package org.example;
 
 public class Statistics {
     private final int maxRequestsPerSeconds;
-    private final int averageRequestsPerSeconds;
+    private final double averageRequestsPerSeconds;
 
     public int getMaxRequestsPerSeconds() {
         return maxRequestsPerSeconds;
     }
 
-    public int getAverageRequestsPerSeconds() {
+    public double getAverageRequestsPerSeconds() {
         return averageRequestsPerSeconds;
     }
 
-    public Statistics(int maxRequestsPerSeconds, int averageRequestsPerSeconds) {
+    public Statistics(int maxRequestsPerSeconds, double averageRequestsPerSeconds) {
         this.maxRequestsPerSeconds = maxRequestsPerSeconds;
         this.averageRequestsPerSeconds = averageRequestsPerSeconds;
     }
